# qa_tmp/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KrishiVaani API",
    "description": "KrishiVaani AI/ML Farming Assistant Platform Backend API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/chatbot/chat": {
      "post": {
        "tags": [
          "LangChain Tool-Calling Conversational Chatbot Agent"
        ],
        "summary": "Chat with KrishiVaani Tool-Calling AI Assistant",
        "description": "Processes farmer queries in natural language, automatically triggers underlying ML models (crop rec, yield prediction, weather advisory, disease detection, satellite NDVI), and returns a formatted conversational reply with structured payloads.",
        "operationId": "chat_with_agent_api_v1_chatbot_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agriculture/chat": {
      "post": {
        "tags": [
          "Agricultural AI Assistant"
        ],
        "summary": "Chat with Agricultural AI Assistant",
        "operationId": "chat_with_agriculture_assistant_api_v1_agriculture_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message",
            "description": "Farmer user prompt or question"
          },
          "farmer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farmer Id",
            "description": "Optional registered farmer ID"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Session ID for conversation memory continuity"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District",
            "description": "District location context"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State",
            "description": "State location context"
          },
          "soil_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Type",
            "description": "Soil type context (e.g. Alluvial, Black, Red)"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "Language code (en, hi, or)",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "example": {
          "district": "Cuttack",
          "farmer_id": "f_123456",
          "language": "en",
          "message": "What crop should I grow in Cuttack with alluvial soil and high rainfall?",
          "session_id": "sess_abc123",
          "soil_type": "Alluvial",
          "state": "Odisha"
        }
      },
      "ChatResponse": {
        "properties": {
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "tools_invoked": {
            "items": {
              "$ref": "#/components/schemas/ToolInvocationResult"
            },
            "type": "array",
            "title": "Tools Invoked",
            "default": []
          },
          "structured_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structured Payload"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "reply",
          "language"
        ],
        "title": "ChatResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ToolInvocationResult": {
        "properties": {
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "input_args": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Args"
          },
          "output_summary": {
            "type": "string",
            "title": "Output Summary"
          }
        },
        "type": "object",
        "required": [
          "tool_name",
          "input_args",
          "output_summary"
        ],
        "title": "ToolInvocationResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}